requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
requests
email-validator
imap-tools